pyproject.toml
[project]
name = "electric"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "poethepoet>=0.40.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.11",
    "ty>=0.0.11",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/**/*.py" = ["S101"] # Allow assert in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ty]
# ty configuration - uses sensible defaults
# See: https://docs.astral.sh/ty/

[tool.poe.tasks]
test = "pytest -v"
test-cov = "pytest --cov=app --cov-report=html --cov-report=term"
lint = "ruff check ."
lint-fix = "ruff check --fix ."
format = "ruff format ."
format-check = "ruff format --check ."
typecheck = "ty check app"
pre-commit = "pre-commit run --all-files"
pre-commit-install = "pre-commit install"
check = ["lint", "format-check", "typecheck", "test"]
